Ensure correct AR version gets loaded in tests
Previously, ginger was falling back to a normal require when
encountering "active_record/version". Subsequently, depending on your
system gems, it would fail with an 'already activated' error or always load
rspec 2 (it's unclear why the different behavior).

Whatever the reason, this is a sure fix. Will load AR proper, then the
version (needed for Rails < 2.2.3)