Add ability to set the TableDefinition class.
The table definition class can be set via passing
in `:table_definition_class_name` as a connection
parameter.

This is specifically useful when the default 
table definition does not support the column types
in the schema.rb. Specifically, passing in
`ActiveRecord::ConnectionAdapters::PostgreSQL::TableDefinition`
allows column types such as `json` to exist in the
schema.